+++ tests/test_client_device.py
@@ -46,8 +46,8 @@ if ENABLE_COMMLOG:
     commlog.defaultLogger = commLogger
 
 CLIENT_VALIDATE = True
-SET_TIMEOUT = 10  # longer timeout than usually needed, but jenkins jobs frequently failed with 3 seconds timeout
-NOTIFICATION_TIMEOUT = 5  # also jenkins related value
+SET_TIMEOUT = 10  # longer timeout than usually needed
+NOTIFICATION_TIMEOUT = 5
 
 
 def mklogger(logFolder=None):
@@ -601,7 +601,7 @@ class Test_Client_SomeDevice(unittest.TestCase):
             firstValue = 12
             myPhysicalConnector = pmtypes.PhysicalConnectorInfo([pmtypes.LocalizedText('ABC')], 1)
             now = time.time()
-            with sdcDevice.mdib.mdibUpdateTransaction(setDeterminationTime=False) as mgr:
+            with sdcDevice.mdib.mdibUpdateTransaction() as mgr:
                 st = mgr.getMetricState(descriptorHandle)
                 if st.metricValue is None:
                     st.mkMetricValue()
@@ -706,6 +706,7 @@ class Test_Client_SomeDevice(unittest.TestCase):
                     st.ActivationState = _activationState
                     st.ActualPriority = _actualPriority
                     st.Presence = _presence
+                    st.DeterminationTime = time.time()
                 coll.result(timeout=NOTIFICATION_TIMEOUT)
                 clientStateContainer = cl_mdib.states.descriptorHandle.getOne(
                     descriptorHandle)  # this shall be updated by notification
@@ -1389,11 +1390,10 @@ class Test_Client_SomeDevice(unittest.TestCase):
         metric_descriptor_handle = '0x34F00100'
         alert_descriptor_handle = '0xD3C00100'
         component_descriptor_handle = '2.1.1'
-        context_descriptor_handle = 'LC.mds0'
         operationalstate_descriptor_handle = 'SVO.42.2.1.1.2.0-6'
         waveform_descriptor_handle = '0x34F05505'
         list_of_handles = [metric_descriptor_handle, alert_descriptor_handle, component_descriptor_handle,
-                           context_descriptor_handle, operationalstate_descriptor_handle, waveform_descriptor_handle]
+                           operationalstate_descriptor_handle, waveform_descriptor_handle]
         self.sdcDevice_Final.mdib._waveform_source._waveform_generators = {}  # stop sending waveforms for this test
         for sdcClient, sdcDevice in self._all_cl_dev:
             # set value of a metric
@@ -1420,7 +1420,6 @@ class Test_Client_SomeDevice(unittest.TestCase):
             metric_coll = observableproperties.SingleValueCollector(sdcClient.mdib, 'metricsByHandle')
             alert_coll = observableproperties.SingleValueCollector(sdcClient.mdib, 'alertByHandle')
             component_coll = observableproperties.SingleValueCollector(sdcClient.mdib, 'componentByHandle')
-            context_coll = observableproperties.SingleValueCollector(sdcClient.mdib, 'contextByHandle')
             operation_coll = observableproperties.SingleValueCollector(sdcClient.mdib, 'operationByHandle')
             waveform_coll = observableproperties.SingleValueCollector(sdcClient.mdib, 'waveformByHandle')
 
@@ -1441,7 +1440,6 @@ class Test_Client_SomeDevice(unittest.TestCase):
             metrics_dict = metric_coll.result(timeout=NOTIFICATION_TIMEOUT)
             alert_dict = alert_coll.result(timeout=NOTIFICATION_TIMEOUT)
             component_dict = component_coll.result(timeout=NOTIFICATION_TIMEOUT)
-            context_dict = context_coll.result(timeout=NOTIFICATION_TIMEOUT)
             operation_dict = operation_coll.result(timeout=NOTIFICATION_TIMEOUT)
             waveform_dict = waveform_coll.result(timeout=NOTIFICATION_TIMEOUT)
 
@@ -1467,12 +1465,6 @@ class Test_Client_SomeDevice(unittest.TestCase):
                 self.assertEqual(len(values_dict), 1)
                 self.assertTrue(handle in values_dict)
                 self.assertEqual(descriptor_versions[handle] + 1, values_dict[handle].DescriptorVersion)
-
-            # verify context state update (they are different from others because they use Handle as key
-            # instead of DescriptorHandle
-            for handle, state in context_dict.items():
-                self.assertEqual(handle, state.Handle)  # verify that context states dict has Handle as key
-                self.assertEqual(context_descriptor_handle, state.descriptorHandle)
 
             # verify that client got updates
             descriptorContainer = clientMdib.descriptions.handle.getOne(metric_descriptor_handle)
@@ -1511,10 +1503,11 @@ class Test_Client_SomeDevice(unittest.TestCase):
             cl_descriptorContainer = clientMdib.descriptions.handle.getOne(new_handle, allowNone=True)
             self.assertIsNone(cl_descriptorContainer)
 
-    def test_AlertConditionModification_Final(self):
-        self._test_AlertConditionModification(self.sdcClient_Final, self.sdcDevice_Final)
+    def test_AlertConditionModification(self):
 
-    def _test_AlertConditionModification(self, sdcClient, sdcDevice):
+        sdcClient = self.sdcClient_Final
+        sdcDevice = self.sdcDevice_Final
+        
         alertDescriptorHandle = '0xD3C00100'
         limitAlertDescriptorHandle = '0xD3C00108'
 
@@ -1590,78 +1583,6 @@ class Test_Client_SomeDevice(unittest.TestCase):
             self.assertEqual(cl_mdsDescriptor.ModelNumber, '1.09')
             self.assertEqual(cl_mdsDescriptor.Manufacturer[-1].text, u'Draeger GmbH')
 
-    def test_remove_add_mds(self):
-        for sdcClient, sdcDevice in self._all_cl_dev:
-            full_mdib, mdib_version_group = copy.deepcopy(sdcDevice.mdib.reconstructMdibWithContextStates())
-            sdcDevice._runRtSampleThread = False
-            time.sleep(0.1)
-            clientMdib = ClientMdibContainer(sdcClient)
-            clientMdib.initMdib()
-            dev_descriptor_count1 = len(sdcDevice.mdib.descriptions.objects)
-            dev_state_count1 = len(sdcDevice.mdib.states.objects)
-            dev_state_count1_handles = set([s.descriptorHandle for s in sdcDevice.mdib.states.objects])
-            descr_handles = list(sdcDevice.mdib.descriptions.handle.keys())
-            state_descriptorHandles = list(sdcDevice.mdib.states.descriptorHandle.keys())
-            contextState_handles = list(sdcDevice.mdib.contextStates.handle.keys())
-            coll = observableproperties.SingleValueCollector(sdcClient, 'descriptionModificationReport')
-            with sdcDevice.mdib.mdibUpdateTransaction() as mgr:
-                mdsDescriptor = sdcDevice.mdib.descriptions.NODETYPE.getOne(namespaces.domTag('MdsDescriptor'))
-                mgr.removeDescriptor(mdsDescriptor.handle)
-            coll.result(timeout=NOTIFICATION_TIMEOUT)
-            # verify that all state versions were saved
-            descr_handles_lookup1 = copy.copy(sdcDevice.mdib.descriptions.handle_version_lookup)
-            state_descriptorHandles_lookup1 = copy.copy(sdcDevice.mdib.states.handle_version_lookup)
-            contextState_descriptorHandles_lookup1 = copy.copy(sdcDevice.mdib.contextStates.handle_version_lookup)
-            for h in descr_handles:
-                self.assertTrue(h in descr_handles_lookup1)
-            for h in state_descriptorHandles:
-                self.assertTrue(h in state_descriptorHandles_lookup1)
-            for h in contextState_handles:
-                self.assertTrue(h in contextState_descriptorHandles_lookup1)
-
-            # verify that client mdib has same number of objects as device mdib
-            dev_descriptor_count2 = len(sdcDevice.mdib.descriptions.objects)
-            dev_state_count2 = len(sdcDevice.mdib.states.objects)
-            dev_state_count2_handles = set([s.descriptorHandle for s in sdcDevice.mdib.states.objects])
-            cl_descriptor_count2 = len(clientMdib.descriptions.objects)
-            cl_state_count2 = len(clientMdib.states.objects)
-            self.assertTrue(dev_descriptor_count2 < dev_descriptor_count1)
-            self.assertEqual(dev_descriptor_count2, 0)
-            self.assertEqual(dev_descriptor_count2, cl_descriptor_count2)
-            self.assertEqual(dev_state_count2, cl_state_count2)
-
-            # now add mds again:
-            with sdcDevice.mdib.mdibUpdateTransaction() as mgr:
-                sdcDevice.mdib.addMdsNode(full_mdib)
-            time.sleep(5)  # difficult to say which observable is updated as the last one, therefore sleep
-            # verify that all objects have a state version at least incremented by one
-            for handle, version in descr_handles_lookup1.items():
-                obj = sdcDevice.mdib.descriptions.handle.getOne(handle)
-                self.assertGreater(obj.DescriptorVersion, version)
-            for handle, version in state_descriptorHandles_lookup1.items():
-                obj = sdcDevice.mdib.states.descriptorHandle.getOne(handle, allowNone=True)
-                if obj:
-                    self.assertGreater(obj.StateVersion, version,
-                                       msg='state {}: {} not greater than {}'.format(obj, obj.StateVersion, version))
-            for handle, version in contextState_descriptorHandles_lookup1.items():
-                obj = sdcDevice.mdib.contextStates.handle.getOne(handle)
-                print('checking object {} state={} expected={}'.format(obj, obj.StateVersion, version + 1))
-                self.assertGreater(obj.StateVersion, version,
-                                   msg='state {}: {} not greater than {}'.format(obj, obj.StateVersion, version + 1))
-
-            dev_descriptor_count3 = len(sdcDevice.mdib.descriptions.objects)
-            dev_state_count3 = len(sdcDevice.mdib.states.objects)
-            dev_state_count3_handles = set([s.descriptorHandle for s in sdcDevice.mdib.states.objects])
-            cl_descriptor_count3 = len(clientMdib.descriptions.objects)
-            cl_state_count3 = len(clientMdib.states.objects)
-            self.assertEqual(dev_descriptor_count3, dev_descriptor_count1)
-            self.assertEqual(dev_descriptor_count3, cl_descriptor_count3)
-            if sdcDevice is self.sdcDevice_Final:
-                self.assertEqual(dev_state_count3, dev_state_count1)
-            else:
-                self.assertEqual(dev_state_count3, dev_state_count1 - 1)  # scostate is not sent in draft6
-            self.assertEqual(dev_state_count3, cl_state_count3)
-
     def test_clientmdib_observables(self):
         for sdcClient, sdcDevice in self._all_cl_dev:
             clientMdib = ClientMdibContainer(sdcClient)
@@ -1671,7 +1592,7 @@ class Test_Client_SomeDevice(unittest.TestCase):
                                                              'metricsByHandle')  # wait for the next EpisodicMetricReport
             descriptorHandle = '0x34F00100'
             firstValue = 12
-            with sdcDevice.mdib.mdibUpdateTransaction(setDeterminationTime=False) as mgr:
+            with sdcDevice.mdib.mdibUpdateTransaction() as mgr:
                 st = mgr.getMetricState(descriptorHandle)
                 if st.metricValue is None:
                     st.mkMetricValue()
@@ -1686,7 +1607,7 @@ class Test_Client_SomeDevice(unittest.TestCase):
             coll = observableproperties.SingleValueCollector(clientMdib,
                                                              'alertByHandle')  # wait for the next EpisodicAlertReport
             descriptorHandle = '0xD3C00108'  # an AlertConditionDescriptorHandle
-            with sdcDevice.mdib.mdibUpdateTransaction(setDeterminationTime=False) as mgr:
+            with sdcDevice.mdib.mdibUpdateTransaction() as mgr:
                 st = mgr.getAlertState(descriptorHandle)
                 st.Presence = True
                 st.Rank = 3
@@ -1697,7 +1618,7 @@ class Test_Client_SomeDevice(unittest.TestCase):
 
             coll = observableproperties.SingleValueCollector(clientMdib, 'updatedDescriptorByHandle')
             descriptorHandle = '0x34F00100'
-            with sdcDevice.mdib.mdibUpdateTransaction(setDeterminationTime=False) as mgr:
+            with sdcDevice.mdib.mdibUpdateTransaction() as mgr:
                 descr = mgr.getDescriptor(descriptorHandle)
                 descr.DeterminationPeriod = 42
             data = coll.result(timeout=NOTIFICATION_TIMEOUT)