+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - remove ContainerBase.diff() method
+- only allow one type of report (e.g. DescriptionModificationReport, EpisodicAlertReport ...) per `mdibUpdateTransaction` context
+- `setDeterminationTime` is not provided anymore by the `mdibUpdateTransaction`
+- remove `addMdsNode` from `DeviceMdibContainer`
 
 ## [1.3.3] - 2026-01-06